feat: refactor job storage to job repository, remove unused job storage classes and update related configurations
