+++ core/config/config.yaml
@@ -50,6 +50,8 @@ di_container:
       url: ${JOB_REPOSITORY_URL, "http://localhost:8888"}
       api_key: ${JOB_REPOSITORY_API_KEY, ""}
       workers: ${JOB_REPOSITORY_WORKERS, 10}
+      storage_op_timeout_seconds: ${JOB_STORAGE_OP_TIMEOUT_SECONDS, 60}
+      max_file_size: ${MAX_FILE_SIZE, 10485760}
 
     _device_repository:
       _target_: oqtopus_engine_core.repositories.NullDeviceRepository
@@ -59,16 +61,6 @@ di_container:
       url: ${DEVICE_REPOSITORY_URL, "http://localhost:8888"}
       api_key: ${DEVICE_REPOSITORY_API_KEY, ""}
       workers: ${DEVICE_REPOSITORY_WORKERS, 1}
-
-    # storage configurations
-    _job_storage:
-      _target_: oqtopus_engine_core.storages.NullJobStorage
-
-    job_storage:
-      _target_: oqtopus_engine_core.storages.OqtopusCloudJobStorage
-      workers: ${JOB_STORAGE_WORKERS, 10}
-      storage_op_timeout_seconds: ${JOB_STORAGE_OP_TIMEOUT_SECONDS, 60}
-      max_file_size: ${MAX_FILE_SIZE, 10485760}
 
     # buffer configurations
     buffer:
