+++ core/config/sse_engine_config.yaml
@@ -38,9 +38,6 @@ di_container:
     device_repository:
       _target_: oqtopus_engine_core.repositories.NullDeviceRepository
 
-    job_storage:
-      _target_: oqtopus_engine_core.storages.NullJobStorage
-
     # buffer configurations
     buffer:
       _target_: oqtopus_engine_core.buffers.QueueBuffer
