+++ core/src/oqtopus_engine_core/fetchers/repository_job_fetcher.py
@@ -93,7 +93,7 @@ class RepositoryJobFetcher(JobFetcher):
         try:
             # Create a list of awaitable tasks for downloading job inputs
             job_download_tasks = [
-                self.gctx.job_storage.download_job_input(job) for job in jobs
+                self.gctx.job_repository.download_job_input(job) for job in jobs
             ]
             # Run all download tasks concurrently
             job_download_results = await asyncio.gather(
