+++ core/tests/oqtopus_engine_core/repositories/test_null_job_repository.py
@@ -1,6 +1,10 @@
 import pytest
 
 from oqtopus_engine_core.framework import Job
+from oqtopus_engine_core.interfaces.oqtopus_cloud import (
+    JobsJobInfoUploadPresignedURL,
+    JobsJobInfoUploadPresignedURLFields,
+)
 from oqtopus_engine_core.repositories.null_job_repository import NullJobRepository
 
 
@@ -16,6 +20,7 @@ def _make_job() -> Job:
         simulator_info={},
         mitigation_info={},
         status="running",
+        output_files=[],
     )
 
 
@@ -40,3 +45,31 @@ async def test_get_job_upload_url_returns_placeholder_urls() -> None:
         "job-1/transpile_result",
         "job-1/result",
     ]
+
+
+@pytest.mark.asyncio
+async def test_download_job_input_returns_empty_dict() -> None:
+    repository = NullJobRepository()
+
+    downloaded = await repository.download_job_input(_make_job())
+
+    assert downloaded == {}
+
+
+@pytest.mark.asyncio
+async def test_upload_job_output_does_not_mutate_output_files() -> None:
+    repository = NullJobRepository()
+    job = _make_job()
+    presigned_url = JobsJobInfoUploadPresignedURL(
+        url="null://upload",
+        fields=JobsJobInfoUploadPresignedURLFields(key="job-1/result"),
+    )
+
+    await repository.upload_job_output(
+        job=job,
+        presigned_url=presigned_url,
+        data={"ok": True},
+        arcname_ext=".json",
+    )
+
+    assert job.output_files == []
